Move data-selector to its own folder as it's used by multiple panes
