+++ src/components/data-selector/data-selector.scss
